Exclude alphanumeric identifier from lexical analysis
Lexical convention of C- includes rules about whitespaces.
White space consists of blanks, newlines, and tabs. White space is
ignored except that it must seperate ID's, NUM's, and keywords.
Also ID's and keywords are only consisted of lower-case and upper-
case alphabets, and NUM's are only consisted of numbers, 0~9.
So tokens like below are not valid tokens.
abc123, 123abc, abc123def, ...
So, we can conclude that if token consists of both alphabets and
numbers, the token is not a valid token. Those cases were not
caught before. For example, "abc123" was understood as "abc" and
"123" seperately.