Exclude exceptions for analyzing C-style multiline comments
C- is a subset of C, so if definitions are ambiguous, it is safe
to follow ANSI C or C90 Standards. In those standards, comment
start token can appear inside comments, like the following.
/* Not an error /* */
/* /* Not an error */
But those cases raised an error in lexical analyzer.
Note that nested comments are still not supported, and thus
analyzer raises an error for the following code.
/* /* This is an error */ */